Move signed extensions to a separate module
